requirements.txt
qypi==0.6.0
PyYAML==6.0
packaging==21.3
